toolkit/components/passwordmgr/test/mochitest/test_autocomplete_basic_form_formActionOrigin.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test that logins with non-matching formActionOrigin appear in autocomplete dropdown</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: logins with non-matching formActionOrigin appear in autocomplete dropdown

<script>
var chromeScript = runChecksAfterCommonInit();

addLoginsInParent([window.location.origin, "https://differentFormSubmitURL", null,
                  "dfsu1", "dfsp1", "uname", "pword"]);
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 tests multiple matching logins -->
  <form id="form1" action="https://autocomplete:8888/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <button type="submit">Submit</button>
  </form>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: multiple login autocomplete. **/

const uname = getFormElementByName(1, "uname");
const pword = getFormElementByName(1, "pword");

// Restore the form to the default state.
function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

function spinEventLoop() {
  return Promise.resolve();
}

add_setup(async () => {
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

/* For this testcase, the only login that exists for this origin
 * is one with a different formActionOrigin, so the login will appear
 * in the autocomplete popup.
 */
add_task(async function test_form1_menu_shows_logins_for_different_formActionOrigin() {
  await SimpleTest.promiseFocus(window);
  // Trigger autocomplete popup
  restoreForm();
  const autocompleteItems = await openAutocompletePopup();

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = ["dfsu1"];
  checkAutoCompleteResults(autocompleteItems, expectedMenuItems, window.location.host, "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // first item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  checkLoginForm(uname, "dfsu1", pword, "dfsp1");
});
</script>
</pre>
</body>
</html>
